+++ devfactory/qa/__init__.py
@@ -1,3 +1,0 @@
-from devfactory.qa.runner import QARunner
-
-__all__ = ["QARunner"]